Moved gui and logic into separate packages
