+++ encrypting/__main__.py
@@ -1,0 +1,9 @@
+from encrypting.gui.window import Window
+from PyQt5.QtWidgets import QApplication
+import sys
+
+app = QApplication(sys.argv)
+app.setStyle('Windows')
+window = Window()
+window.show()
+sys.exit(app.exec_())
