+++ encrypting/logic/ecb.py
@@ -1,17 +1,10 @@
-from symkey import SymKey
-import os
-import Crypto
+from encrypting.logic.symkey import SymKey
 from Crypto import Random
 from Crypto.Cipher import AES
-class Ctr():
+class Ecb():
         def __init__(self, sym_key):
             self.key = sym_key.key
-
-            if sym_key.counter == None:
-                self.counter = os.urandom(16)
-            else:
-                self.counter = sym_key.counter
-            self.cipher = AES.new(self.key,AES.MODE_CTR,counter = lambda : self.counter)
+            self.cipher = AES.new(self.key,AES.MODE_ECB)
         
         def encode(self, data):
             return self.cipher.encrypt(data)
@@ -24,5 +17,5 @@ class Ctr():
             if not (key_size % 16 == 0):
                 raise ArithmeticError("key size must be multiple of 16")
             rand =Random.new()
-            return SymKey(rand.read(key_size),counter=os.urandom(16))
+            return SymKey(rand.read(key_size))
